+++ modAionImpl/src/org/aion/zero/impl/blockchain/AionBlockchainImpl.java
@@ -30,9 +30,9 @@ import java.util.Set;
 import java.util.Stack;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
-import org.aion.vm.common.PostExecutionLogic;
-import org.aion.vm.common.PostExecutionWork;
-import org.aion.vm.common.VmFatalException;
+import org.aion.zero.impl.vm.common.PostExecutionLogic;
+import org.aion.zero.impl.vm.common.PostExecutionWork;
+import org.aion.zero.impl.vm.common.VmFatalException;
 import org.aion.base.AccountState;
 import org.aion.base.AionTransaction;
 import org.aion.base.ConstantUtil;
@@ -75,8 +75,8 @@ import org.aion.util.types.AddressUtils;
 import org.aion.util.types.ByteArrayWrapper;
 import org.aion.util.types.Hash256;
 import org.aion.utils.HeapDumper;
-import org.aion.vm.common.BlockCachingContext;
-import org.aion.vm.common.BulkExecutor;
+import org.aion.zero.impl.vm.common.BlockCachingContext;
+import org.aion.zero.impl.vm.common.BulkExecutor;
 import org.aion.zero.impl.config.CfgAion;
 import org.aion.zero.impl.core.energy.AbstractEnergyStrategyLimit;
 import org.aion.zero.impl.core.energy.EnergyStrategies;
