+++ modAionImpl/src/org/aion/zero/impl/blockchain/AionImpl.java
@@ -2,7 +2,7 @@ package org.aion.zero.impl.blockchain;
 
 import java.util.List;
 import java.util.Optional;
-import org.aion.vm.common.VmFatalException;
+import org.aion.zero.impl.vm.common.VmFatalException;
 import org.aion.base.AionTransaction;
 import org.aion.crypto.ECKey;
 import org.aion.crypto.ECKeyFac;
@@ -17,8 +17,8 @@ import org.aion.mcf.db.RepositoryCache;
 import org.aion.types.AionAddress;
 import org.aion.util.types.AddressUtils;
 import org.aion.util.types.ByteArrayWrapper;
-import org.aion.vm.common.BlockCachingContext;
-import org.aion.vm.common.BulkExecutor;
+import org.aion.zero.impl.vm.common.BlockCachingContext;
+import org.aion.zero.impl.vm.common.BulkExecutor;
 import org.aion.zero.impl.SystemExitCodes;
 import org.aion.zero.impl.config.CfgAion;
 import org.aion.zero.impl.db.AionRepositoryImpl;
