+++ modAionImpl/src/org/aion/zero/impl/blockchain/StakingContractHelper.java
@@ -9,11 +9,11 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import org.aion.base.AionTxExecSummary;
-import org.aion.vm.avm.AvmProvider;
-import org.aion.vm.avm.AvmConfigurations;
-import org.aion.vm.avm.AvmTransactionExecutor;
-import org.aion.vm.common.BlockCachingContext;
-import org.aion.vm.common.VmFatalException;
+import org.aion.zero.impl.vm.avm.AvmProvider;
+import org.aion.zero.impl.vm.avm.AvmConfigurations;
+import org.aion.zero.impl.vm.avm.AvmTransactionExecutor;
+import org.aion.zero.impl.vm.common.BlockCachingContext;
+import org.aion.zero.impl.vm.common.VmFatalException;
 import org.aion.avm.stub.AvmVersion;
 import org.aion.avm.stub.IAvmResourceFactory;
 import org.aion.base.AccountState;
