+++ modAionImpl/src/org/aion/zero/impl/cli/Cli.java
@@ -19,9 +19,9 @@ import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import org.aion.avm.stub.AvmVersion;
-import org.aion.vm.avm.AvmConfigurations;
-import org.aion.vm.avm.AvmProvider;
-import org.aion.vm.avm.schedule.AvmVersionSchedule;
+import org.aion.zero.impl.vm.avm.AvmConfigurations;
+import org.aion.zero.impl.vm.avm.AvmProvider;
+import org.aion.zero.impl.vm.avm.schedule.AvmVersionSchedule;
 import org.aion.zero.impl.config.CfgAion;
 import org.aion.zero.impl.keystore.Keystore;
 import org.aion.mcf.config.Cfg;
