+++ modAionImpl/src/org/aion/zero/impl/vm/avm/AvmConfigurations.java
@@ -1,8 +1,8 @@
-package org.aion.vm.avm;
+package org.aion.zero.impl.vm.avm;
 
-import org.aion.vm.avm.schedule.AvmVersionSchedule;
+import org.aion.zero.impl.vm.avm.schedule.AvmVersionSchedule;
 import org.aion.avm.stub.IEnergyRules;
-import org.aion.vm.common.TxNrgRule;
+import org.aion.zero.impl.vm.common.TxNrgRule;
 
 /**
  * The class that holds all details pertaining to how the avm is set up.
