+++ modAionImpl/src/org/aion/zero/impl/vm/avm/AvmTransactionExecutor.java
@@ -1,4 +1,4 @@
-package org.aion.vm.avm;
+package org.aion.zero.impl.vm.avm;
 
 import java.io.IOException;
 import java.math.BigInteger;
@@ -6,9 +6,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
-import org.aion.vm.avm.schedule.AvmVersionSchedule;
-import org.aion.vm.common.PostExecutionWork;
-import org.aion.vm.common.VmFatalException;
+import org.aion.zero.impl.vm.avm.schedule.AvmVersionSchedule;
+import org.aion.zero.impl.vm.common.PostExecutionWork;
+import org.aion.zero.impl.vm.common.VmFatalException;
 import org.aion.avm.stub.AvmExecutionType;
 import org.aion.avm.stub.AvmVersion;
 import org.aion.avm.stub.IAionVirtualMachine;
