+++ modAionImpl/src/org/aion/zero/impl/vm/common/PostExecutionWork.java
@@ -1,4 +1,4 @@
-package org.aion.vm.common;
+package org.aion.zero.impl.vm.common;
 
 import org.aion.base.AccountState;
 import org.aion.base.AionTransaction;
