+++ modAionImpl/src/org/aion/zero/impl/vm/fvm/ExternalCapabilitiesForFvm.java
@@ -1,4 +1,4 @@
-package org.aion.vm.fvm;
+package org.aion.zero.impl.vm.fvm;
 
 import java.math.BigInteger;
 import org.aion.base.TxUtil;
