+++ modAionImpl/test/org/aion/zero/impl/vm/ContractIntegTest.java
@@ -18,7 +18,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import org.aion.vm.common.VmFatalException;
+import org.aion.zero.impl.vm.common.VmFatalException;
 import org.aion.avm.stub.AvmVersion;
 import org.aion.avm.stub.IAvmResourceFactory;
 import org.aion.avm.stub.IContractFactory.AvmContract;
@@ -42,8 +42,8 @@ import org.aion.types.AionAddress;
 import org.aion.util.bytes.ByteUtil;
 import org.aion.util.conversions.Hex;
 import org.aion.util.types.AddressUtils;
-import org.aion.vm.common.BlockCachingContext;
-import org.aion.vm.common.BulkExecutor;
+import org.aion.zero.impl.vm.common.BlockCachingContext;
+import org.aion.zero.impl.vm.common.BulkExecutor;
 import org.aion.zero.impl.blockchain.StandaloneBlockchain;
 import org.aion.zero.impl.blockchain.StandaloneBlockchain.Builder;
 import org.aion.zero.impl.db.AionRepositoryCache;
