+++ modAionImpl/test/org/aion/zero/impl/vm/SolidityTypeTest.java
@@ -30,7 +30,7 @@ import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import org.aion.vm.common.VmFatalException;
+import org.aion.zero.impl.vm.common.VmFatalException;
 import org.aion.base.AionTransaction;
 import org.aion.base.TransactionTypes;
 import org.aion.crypto.ECKey;
@@ -56,8 +56,8 @@ import org.aion.solidity.SolidityType.StringType;
 import org.aion.types.AionAddress;
 import org.aion.util.bytes.ByteUtil;
 import org.aion.util.conversions.Hex;
-import org.aion.vm.common.BlockCachingContext;
-import org.aion.vm.common.BulkExecutor;
+import org.aion.zero.impl.vm.common.BlockCachingContext;
+import org.aion.zero.impl.vm.common.BulkExecutor;
 import org.aion.zero.impl.blockchain.StandaloneBlockchain;
 import org.aion.zero.impl.db.AionRepositoryImpl;
 import org.aion.zero.impl.types.AionBlock;
