+++ modAionImpl/test/org/aion/zero/impl/vm/TransactionExecutorTest.java
@@ -32,7 +32,7 @@ import java.io.IOException;
 import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.Collections;
-import org.aion.vm.common.VmFatalException;
+import org.aion.zero.impl.vm.common.VmFatalException;
 import org.aion.base.AionTransaction;
 import org.aion.base.TransactionTypes;
 import org.aion.base.TxUtil;
@@ -44,8 +44,8 @@ import org.aion.mcf.db.RepositoryCache;
 import org.aion.util.types.DataWord;
 import org.aion.types.AionAddress;
 import org.aion.util.conversions.Hex;
-import org.aion.vm.common.BlockCachingContext;
-import org.aion.vm.common.BulkExecutor;
+import org.aion.zero.impl.vm.common.BlockCachingContext;
+import org.aion.zero.impl.vm.common.BulkExecutor;
 import org.aion.zero.impl.types.BlockContext;
 import org.aion.zero.impl.blockchain.StandaloneBlockchain;
 import org.aion.zero.impl.blockchain.StandaloneBlockchain.Builder;
