+++ modApiServer/build.gradle
@@ -14,9 +14,6 @@ sourceSets {
 }
 
 dependencies {
-    //compile 'network.aion:util4j:0.4.0'
-    //compile 'network.aion:log4j:0.4.0'
-    //compile 'network.aion:crypto4j:0.4.0'
     compile project(':modUtil')
     compile project(':modLogger')
     compile project(':modCrypto')
@@ -50,7 +47,6 @@ dependencies {
     testCompile 'junit:junit:4.12'
     testCompile 'org.hamcrest:hamcrest-all:1.3'
     testCompile "org.mockito:mockito-core:2.23.0"
-    testCompile project(':modVM')
 }
 
 test {
