+++ modApiServer/test/org/aion/api/server/AvmTestConfig.java
@@ -1,7 +1,7 @@
 package org.aion.api.server;
 
-import org.aion.vm.avm.AvmConfigurations;
-import org.aion.vm.avm.schedule.AvmVersionSchedule;
+import org.aion.zero.impl.vm.avm.AvmConfigurations;
+import org.aion.zero.impl.vm.avm.schedule.AvmVersionSchedule;
 
 public final class AvmTestConfig {
