+++ modBoot/build.gradle
@@ -1,14 +1,10 @@
 ext.moduleName = 'aion.boot'
 
 dependencies {
-    //compile 'network.aion:util4j:0.4.0'
-    //compile 'network.aion:log4j:0.4.0'
-    //compile 'network.aion:crypto4j:0.4.0'
     compile project(':modUtil')
     compile project(':modLogger')
     compile project(':modCrypto')
     compile project(':modAvmStub')
-    compile project(':modVM')
 
     compile project(':modApiServer')
     compile project(':modAionImpl')
