+++ modBoot/src/org/aion/Aion.java
@@ -33,19 +33,16 @@ import org.aion.api.server.pb.ApiAion0;
 import org.aion.api.server.pb.IHdlr;
 import org.aion.api.server.zmq.HdlrZmq;
 import org.aion.api.server.zmq.ProtocolProcessor;
-import org.aion.vm.avm.AvmProvider;
-import org.aion.vm.avm.schedule.AvmVersionSchedule;
-import org.aion.vm.avm.AvmConfigurations;
+import org.aion.zero.impl.vm.avm.AvmProvider;
+import org.aion.zero.impl.vm.avm.schedule.AvmVersionSchedule;
+import org.aion.zero.impl.vm.avm.AvmConfigurations;
 import org.aion.avm.stub.AvmVersion;
-import org.aion.avm.stub.IEnergyRules;
-import org.aion.avm.stub.IEnergyRules.TransactionType;
 import org.aion.crypto.ECKeyFac;
 import org.aion.crypto.HashUtil;
 import org.aion.equihash.EquihashMiner;
 import org.aion.evtmgr.EventMgrModule;
 import org.aion.log.AionLoggerFactory;
 import org.aion.log.LogEnum;
-import org.aion.vm.common.TxNrgRule;
 import org.aion.zero.impl.keystore.Keystore;
 import org.aion.mcf.config.CfgApiRpc;
 import org.aion.mcf.config.CfgApiZmq;
