+++ settings.gradle
@@ -11,7 +11,6 @@ include 'modAionImpl',
     'modDbImpl',
     'modMcf',
     'modPrecompiled',
-    'modVM',
     'modApiServer',
     'modBoot',
     'modAvmVersion2',
